# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRM Nekretnine API",
        "description": "API dokumentacija za CRM sistem nekretnina",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pregledi": {
            "get": {
                "tags": [
                    "Pregledi"
                ],
                "summary": "Lista pregleda",
                "operationId": "c0c7a15891f04be9a6e2e63dc8f816de",
                "responses": {
                    "200": {
                        "description": "Lista pregleda"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pregledi"
                ],
                "summary": "Kreiranje pregleda",
                "operationId": "6cb402b5820ff5de9b2f6e04b7e325c3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kupac_id",
                                    "nekretnina_id",
                                    "datum"
                                ],
                                "properties": {
                                    "kupac_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nekretnina_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "datum": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-02-25"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "zakazan"
                                    },
                                    "napomena": {
                                        "type": "string",
                                        "example": "Kupac želi obilazak popodne."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pregled kreiran"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pregledi/{id}": {
            "get": {
                "tags": [
                    "Pregledi"
                ],
                "summary": "Prikaz pregleda po ID-u",
                "operationId": "38b85547538f306d00fe37ada7080d88",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pregleda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalji pregleda"
                    },
                    "404": {
                        "description": "Pregled nije pronađen"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pregledi"
                ],
                "summary": "Izmena pregleda",
                "operationId": "c8f0b01e67f3097783ef167761e9dcfd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pregleda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "kupac_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "nekretnina_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "datum": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-02-26"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "odrzan"
                                    },
                                    "napomena": {
                                        "type": "string",
                                        "example": "Pregled uspešno obavljen."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pregled izmenjen"
                    },
                    "404": {
                        "description": "Pregled nije pronađen"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pregledi"
                ],
                "summary": "Brisanje pregleda",
                "operationId": "ca7d1a776098e10d3471dda83095c792",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pregleda",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pregled obrisan"
                    },
                    "404": {
                        "description": "Pregled nije pronađen"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/pregledi/search": {
            "get": {
                "tags": [
                    "Pregledi"
                ],
                "summary": "Pretraga pregleda",
                "operationId": "5017a051e3be3884b8919a8f4d3cbabf",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Tekst pretrage (status/napomena)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "zakazan"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Broj strane",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Broj stavki po strani",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter po statusu",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "zakazan"
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Polje za sortiranje",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "datum",
                                "status",
                                "kupac_id",
                                "nekretnina_id"
                            ]
                        },
                        "example": "datum"
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Smer sortiranja",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        },
                        "example": "desc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginirani rezultat pretrage"
                    },
                    "422": {
                        "description": "Validaciona greška"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}